% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpa_launch.R
\name{qpa_launch}
\alias{qpa_launch}
\title{Launch a QPA Tutorial in the Browser}
\usage{
qpa_launch(tutorial)
}
\arguments{
\item{tutorial}{A number (1-19) or a name matching part of the tutorial
folder name (e.g., "bivariateCat", "regression", "logit").}
}
\description{
Opens a QPA tutorial in your default web browser. This is the recommended
way to run tutorials if you are using assistive technology or prefer working
in a browser rather than RStudio's internal viewer.
}
\examples{
\dontrun{
qpa_launch(1)             # R Basics Part 1
qpa_launch(7)             # Bivariate Description: Categorical
qpa_launch("bivariateCat") # same as above, by name
qpa_launch("regression")  # matches first tutorial with "regression" in name
}

}
