% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpa_launch.R
\name{qpa_list}
\alias{qpa_list}
\title{List Available QPA Tutorials}
\usage{
qpa_list()
}
\description{
Prints a numbered list of all available tutorials with their titles.
}
\examples{
\dontrun{
qpa_list()
}

}
