% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_formats.R
\name{qpa_tutorial}
\alias{qpa_tutorial}
\title{qpa tutorial output format (learnr + shared header)}
\usage{
qpa_tutorial(...)
}
\description{
Ensures shared inst/tutorials/shared/tutorial_includes.html is injected
during prerender, while also working in development mode.
}
